requirements.txt
# BrainClaw Plugin Python Dependencies
# 
# INSTALLATION:
# This plugin includes a bundled Python backend (python/openclaw_memory/).
# For production, simply install the dependencies listed below.
#
# For development, you can use the monorepo's openclaw-memory package:
#   pip install -e ../openclaw-memory

# Core dependencies
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
weaviate-client>=4.4.0
neo4j>=5.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
structlog>=24.0.0
prometheus-client>=0.19.0


# LLM Integration (optional - uncomment as needed)
# ollama>=0.1.0    # For Ollama embeddings
# openai>=1.0.0    # For OpenAI embeddings